Changed ArticleController to return ResponseEntity<>

+++ src/main/java/com/tech1/tech1test/services/UserServiceImpl.java
@@ -58,7 +58,10 @@ public class UserServiceImpl implements UserService{
     }
 
     public void delete(User user) {
-        userRepo.delete(user);
+        Optional<User> optionalUser = userRepo.findById(user.getId());
+        if (optionalUser.isPresent()) {
+            userRepo.delete(user);
+        }
     }
 
     public List<String> getUserNamesByArticlesCount(Integer count) {
